requirements.txt
discord.py
python-dotenv
psycopg2-binary
libsql-experimental
flask